+++ UnitTests/DataTests.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using NUnit.Framework;
-using NamedPipeWrapper;
+using ricaun.NamedPipeWrapper;
 using log4net.Appender;
 using log4net.Config;
 using log4net.Layout;