+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.8.0] / 2024-09-26
 ### Features
 - Support `IJsonService` to use `NewtonsoftJsonService` if available.
+### Updated
+- Update namespace to `ricaun.NamedPipeWrapper`.
+- Update package to `ricaun.NamedPipeWrapper.Json`.
 ### Removed
 - Remove `Newtonsoft.Json` reference.